pyproject.toml
[tool.poetry]
name = "reddit_to_shorts"
version = "0.1.0"
description = "Contains all application modules"
authors = ["Vaclav Honzik <vaclavhonzik98@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
praw = "^7.6.0"
fastapi = "^0.78.0"
python-dotenv = "^0.20.0"
pymongo = "^4.1.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
autopep8 = "^1.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
